Refactor deploymentStack config to lax input reqs
Currently, the deploymentStack configuration parsing is too strict in
terms of requirements, i.e., `deny-settings-mode` does not exists for
`delete` action.

In addition, the deploymentStack scope can be optionally specified for
`delete` action, however, the Azure Deployment Stack client does not
currently expose a `beginDeleteAtTenantScopeAndWait` similar to
Azure Deployment client, thus, the highest level of deletion is at
`SubscriptionScope` without a specified subscription ID,
which seems to be akin to `TenantScope`.