Make providing template file optional
When working with deploymentStack, providing a template is not
required when using the `delete` action and can cause confusion.

This commit introduces a flag that determines whether or not a
missing template file is valid.
The default value is assumed to be `true` as it seems to have been
the case historically.